smoother templating
